Refactored to separate collision queries from invertedindex.
