+++ src/util/collisionquery.h
@@ -1,0 +1,116 @@
+/*
+Minetest
+Copyright (C) 2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#include <unordered_map>
+#include <vector>
+
+#include "util/invertedindex.h"
+
+class CollisionContext
+{
+public:
+	CollisionContext(aabb3f box, InvertedIndex *index);
+
+	u32 addIndexList(IndexListIterator *index);
+	u32 subtractIndexList(IndexListIterator *index);
+
+	// Get the bitmask of valid faces. For each valid face, put the
+	// distance between that face of the entity and the opposing face
+	// of the collision box.
+	u16 getValidFaces(u32 id, f32 offset[6]) const
+	{
+		std::unordered_map<u32, CollisionContextDetail>::const_iterator  entry = m_active.find(id);
+		if (entry == m_active.end())
+			return 0;
+
+		u16 valid = entry->second.valid_faces;
+		for (u32 f = 0; f < 6; f++)
+			if ((valid & setBitmask[f]) == setBitmask[f])
+				offset[f] = m_face_offset[f] - entry->second.face_offset[f];
+		return valid;
+	}
+
+	static const u16 setBitmask[];
+	static const u16 unsetBitmask[];
+	static const CollisionFace opposingFace[];
+
+protected:
+	f32 m_face_offset[6];
+	std::unordered_map<u32, CollisionContextDetail> m_active;
+	std::vector<u32> collisions;
+};
+
+/* TODO *
+struct IndexScan
+{
+	CollisionFace face;
+	f32 time;
+	// Physics can be incorporated with a function instead of scale.
+	const f32 begin;
+	const f32 timeScale;
+	u32 handle;
+	const i32 increment;
+	const u32 end;
+	AttributeIndex *next;
+
+	IndexScan(CollisionFace face, f32 time, f32 begin, f32 timeScale, u32 handle, i32 increment, u32 end) :
+			face(face), time(time), begin(begin), timeScale(timeScale),
+			handle(handle), increment(increment), end(end) {}
+	IndexScan(const IndexScan &other) = default;
+};
+
+class QueryScan
+{
+	QueryScan(InvertedIndex *index) : m_index(index) {}
+
+	// Scan the offsets in interval [begin, end)
+	bool addScan(CollisionFace face, f32 begin, f32 velocity, f32 dtime);
+
+	// Perform the scan.
+	// You MUST call these function is order.
+	// nextFace() will tell you what face, if any, the next AttributeIndex
+	// involves. COLLISION_FACE_NONE will indicate the scan is done.
+	CollisionFace nextFace()
+	{
+		if (m_scans.empty())
+			return COLLISION_FACE_NONE;
+		if (!heap)
+			std::make_heap(m_scans->begin(), m_scans->end(), compare);
+		return m_index->front().face;
+	}
+
+	// If nextFace() did not return COLLISION_FACE_NONE, optionally call
+	// nextTime() to get the specific 
+	f32 nextTime() const
+	{
+		return m_index->front().time;
+	}
+
+	// If nextFace() did not return COLLISION_FACE_NONE, call next() to
+	// get the AttributeIndex. nextTime() should not be called after this
+	// until nextFace() is called.
+	AttributeIndex *next();
+
+protected:
+	static bool compare(IndexScan a, IndexScan b) { return a.time < b.time; }
+	InvertedIndex *m_index;
+	std::vector<IndexScan> m_scans;
+	bool m_heap;
+}
+* TODO */
